requirements.txt
absl-py==0.15.0
addict==2.4.0
astunparse==1.6.3
bleach==1.5.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2022.6.15
charset-normalizer==2.0.12
clang==5.0
cycler==0.11.0
dataclasses==0.8
decorator==4.1.2
entrypoints==0.2.3
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.0
h5py==3.1.0
html5lib==0.9999999
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==4.6.1
ipython==6.1.0
ipython_genutils==0.2.0
ipywidgets==6.0.0
jedi==0.10.2
Jinja2==2.9.6
joblib==1.1.0
jsonschema==2.6.0
jupyter_client==5.1.0
jupyter_core==4.3.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.7
MarkupSafe==1.0
matplotlib==3.3.4
mistune==0.7.4
nbconvert==5.2.1
nbformat==4.4.0
notebook==5.0.0
numpy==1.19.5
oauthlib==3.2.0
open3d==0.11.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
pandas==1.1.5
pandocfilters==1.4.2
pexpect==4.2.1
pickleshare==0.7.4
Pillow==8.4.0
plyfile==0.7.4
prompt-toolkit==1.0.15
protobuf==4.21.0
ptyprocess==0.5.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.2.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
pyzmq==16.0.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==0.24.2
scipy==1.5.4
simplegeneric==0.8.1
six==1.15.0
sklearn==0.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.6
testpath==0.3
threadpoolctl==3.1.0
torch==1.10.2+cpu
torchaudio==0.10.2+cpu
torchvision==0.11.3+cpu
tornado==4.5.2
tqdm==4.64.0
traitlets==4.3.2
typing-extensions==3.7.4.3
urllib3==1.26.11
wcwidth==0.1.7
Werkzeug==2.0.3
widgetsnbextension==3.0.2
wrapt==1.12.1
zipp==3.6.0
